+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 2.14 branch
 ### 2.14.7
 * you can choose whether the banners should be displayed with a thumbnail or
-    not. This allows you to view animated gif unchanged.
+    not. This allows you to view animated gif unchanged;
+* banners have `www` virtual field.
 
 ### 2.14.6
 * to generate thumbnails, uses the `fit()` method instead of `crop()`.